Example of Djikstra
A Djikstra that takes (surprisingly) 3 CSV files as input:
-nodes
-links
-nodes but with different labels? I don't remember why I did this
It gives out Djikstra answer to shortest paths.

I did some work to connect to a postgresql database. If I find it, I
will upload it, then it will be easier to use postgis with the pgRouting
to make the shortestpath compution.